requirements.txt
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
flasgger==0.9.5
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
mistune==2.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
python-decouple==3.5
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
validators==0.18.2
Werkzeug==2.0.2
zipp==3.7.0
